+++ core/src/transport/memory.rs
@@ -62,9 +62,8 @@ impl Hub {
             port
         } else {
             loop {
-                let port = match NonZeroU64::new(rand::random()) {
-                    Some(p) => p,
-                    None => continue,
+                let Some(port) = NonZeroU64::new(rand::random()) else {
+                    continue;
                 };
                 if !hub.contains_key(&port) {
                     break port;
@@ -184,16 +183,12 @@ impl Transport for MemoryTransport {
         id: ListenerId,
         addr: Multiaddr,
     ) -> Result<(), TransportError<Self::Error>> {
-        let port = if let Ok(port) = parse_memory_addr(&addr) {
-            port
-        } else {
-            return Err(TransportError::MultiaddrNotSupported(addr));
-        };
+        let port =
+            parse_memory_addr(&addr).map_err(|_| TransportError::MultiaddrNotSupported(addr))?;
 
-        let (rx, port) = match HUB.register_port(port) {
-            Some((rx, port)) => (rx, port),
-            None => return Err(TransportError::Other(MemoryTransportError::Unreachable)),
-        };
+        let (rx, port) = HUB
+            .register_port(port)
+            .ok_or(TransportError::Other(MemoryTransportError::Unreachable))?;
 
         let listener = Listener {
             id,