+++ protocols/identify/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.44.1
+
+- Ensure `Multiaddr` handled and returned by `Behaviour` are `/p2p` terminated.
+  See [PR 4596](https://github.com/libp2p/rust-libp2p/pull/4596).
+
 ## 0.44.0
 
 - Add `Info` to the `libp2p-identify::Event::Pushed` to report pushed info.